Refac : Layout Components styled 분리
