+++ src/components/layouts/CardGrid/CardGrid.jsx
@@ -1,8 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
-import { Link } from "react-router-dom";
+import { Layout, CardItemLink } from "./CardGrid.styled";
 import { CardResponsive } from "@/components";
 
 const THEME = {
@@ -33,20 +32,6 @@ CardGrid.defaultProps = {
   theme: THEME.LIGHT,
 };
 
-export default CardGrid;
-
-const Layout = styled.div`
-  display: grid;
-  width: 100%;
-  max-width: 940px;
-  grid-template-columns: repeat(auto-fill, minmax(300px, auto));
-  gap: 20px;
-`;
-
-const CardItemLink = styled(Link)`
-  text-decoration: none;
-`;
-
 const cardData = new Array(3).fill({
   contents: {
     title: "",
@@ -57,3 +42,5 @@ const cardData = new Array(3).fill({
   badges: ["", "", ""],
   id: "",
 });
+
+export default CardGrid;
