+++ src/components/layouts/CardSmallSlider/CardSmallSlider.jsx
@@ -1,12 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
-
-import { Link } from "react-router-dom";
 
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 
+import { Layout, CardItemLink } from "./CardSmallSlider.styled";
 import { CardSmall, CardSmallCreate } from "@/components";
 
 const CardSmallSlider = ({ createMsg, data, theme, isLoading, ...props }) => {
@@ -52,16 +50,3 @@ const cardData = new Array(6).fill({
   badges: ["", "", ""],
   id: "",
 });
-
-const Layout = styled.div`
-  width: 100%;
-
-  .swiper-slide {
-    width: 220px !important;
-    flex-shrink: 0;
-  }
-`;
-
-const CardItemLink = styled(Link)`
-  text-decoration: none;
-`;
