+++ src/components/layouts/CommentList/CommentList.styled.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Layout = styled.div`
+  padding: 2rem 0;
+`;
+
+export const CommentBox = styled.div`
+  padding: 1rem 0;
+`;
