+++ src/components/layouts/FeedGrid/FeedGrid.jsx
@@ -1,11 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import styled from "styled-components";
 
-import { Link } from "react-router-dom";
-
+import { Layout, FeedItemLink } from "./FeedGrid.styled";
 import { FeedItem } from "@/components";
-import { colors } from "@/_shared";
 
 const FeedGrid = ({ data, isLoading, theme, ...props }) => {
   const feeds = isLoading ? loadingData : data;
@@ -41,26 +38,3 @@ const loadingData = new Array(6).fill({
   commentcount: "",
   viewcount: "",
 });
-
-const borderColor = {
-  light: colors.gray100,
-  dark: colors.gray900,
-};
-
-const Layout = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 32px;
-
-  width: 100%;
-  margin: auto;
-  padding-top: 32px;
-`;
-
-const FeedItemLink = styled(Link)`
-  width: 100%;
-  padding-bottom: 32px;
-  border-bottom: 1px solid ${(props) => borderColor[props.theme]};
-  text-decoration: none;
-`;
